Rename : utils 파일 생성 및 timeAgo 작업
